Add "Mark as read" button for rows
Now, it is possible to mark the post as read right on the page. Added
a new view controller for that purpose.
Also, changed the browser engine from webkit to chromium. The webKit is
used mostly by Safari, and this is not my target browser engine. Plus,
it is not supported on Fedora. For the same reason, I removed the
dependence of the `test` task on the `playwrightInstall` task.

The `/mark-as-read` endpoint in the PostController is deprecated to be
able to find it easier later on to remove it.