Define Testcontainers image in properties file
When a docker tag used for Testcontainers is defined in an external
properties file, then it is possible to write a regex manager for the
Renovate so that it will scan this file, extract an image name and tag
and will scan it against the docker. This way, the image will eb always
up to date.