Refactor integration test to use direct SQL
Refactor the integration test that cleans up the read posts that are
older than 30 days. Instead of mocking time and making HTTP calls to
mark the post as read, insert them into the DB directly and set the time
the post is read on the DB level. It's quicker and easier to manage in
the test.