Add controller to delete old unread posts (#139)
* Add date_read column to the 'posts' table

Set the timestamp when the post is read. It will be later used in the
functionality to delete read posts older than X number of days.

* Add controller to clean up read old posts

Delete all posts that were read more than 30 days ago.

* Provide clock bean when mark post as read

* Use MockMvcTester for existing test

* Add integration test for cleanup controller

* Refactor integration test to use direct SQL

Refactor the integration test that cleans up the read posts that are
older than 30 days. Instead of mocking time and making HTTP calls to
mark the post as read, insert them into the DB directly and set the time
the post is read on the DB level. It's quicker and easier to manage in
the test.

* Add a cleanup job into the pipeline

* Mock PostService in tests